From vlc 2.1, need to set the telnet password
From vlc 2.1 there no more default password for telnet interface.
Moreover, the telnet interface do not start if there is no password set,
so, we do it on the command line (did not find how to do it with an
environnement variable).
This is quite bad as it leek the password in the list of running processes.
(I don't care as I use the default password admin, but so…)

If i don't find anything, I will probably feelup a bug report on vlc tracker.